+++ HISTORY.md
@@ -91,6 +91,7 @@
 ### 2019-06-25
 * threads: 2 optional formal inputs for rings in MAP, KEEP, FIND, COMBINE: index and source list
 * objects, threads: renamed COMBINE label, switching the inputs
+* objects: took out "relabel" options for COMBINE
 
 ### 2019-06-24
 * removed "such that" from KEEP and FIND block labels